Исправлен тип глобальных переменных + код стайл

+++ Mili/src/BonusMilesService.java
@@ -1,8 +1,8 @@
 public class BonusMilesService {
-    static final long BONUS_FACTOR = 1;
-    static final long BONUS_COST = 20;
+    static final int BONUS_FACTOR = 1;
+    static final int BONUS_COST = 20;
 
     public int calculate(int ticketCost) {
-        return (int) (ticketCost/BONUS_COST*BONUS_FACTOR);
+        return (ticketCost / BONUS_COST * BONUS_FACTOR);
     }
 }
